review: gate the switch-back warmth hint, prune dead forecast fields, tighten test realism
Adopted from Slate Lime Moose's review:

- A switch-back prior now needs an exact wire-API match (cache identity is
  provider+id+api everywhere else; a pre-api snapshot no longer hints warmth)
  and an uncompacted path since the prior call (pathContainsCompaction is the
  same gate lineage resolution already uses). A density anchor still matches
  leniently: the tokenizer is the model's, not the wire's.
- Pruned droppedReasoningChars, imageCount and messageCount from
  HistoryForecast: no production consumers; tests were keeping them alive.
  The contract pin for cross-model drops is now an assertMirror against a
  fully foreign target, which checks the real transform instead of the
  mirror's own bookkeeping.
- Event-flow tests now drive model_select itself (flip on select, revive on
  select-back, no notice from selection alone) and moved to
  model-switch-events.test.ts to keep the suite inside its context budget.
- Codex snapshots in fixtures carry band windows, not anthropic contract
  TTLs a production session cannot produce.
- docs/pi-cachemire.md documents the hedge and both gates.

Declined (with reasons in the PR thread): removing the forecast engine, the
A\u2192B\u2192A branch, and the contract suite.
