+++ extensions/_lib/forecast.ts
@@ -50,12 +50,8 @@ export interface HistoryForecast {
   textChars: number;
   /** Same-model reasoning replayed as-is (encrypted payloads, or readable text). */
   keptReasoningChars: number;
-  /** Cross-model reasoning the target will never see (encrypted/redacted payloads). */
-  droppedReasoningChars: number;
-  imageCount: number;
   /** Images at pi's own flat per-image char convention (compaction's estimate). */
   imageChars: number;
-  messageCount: number;
 }
 
 // pi's compaction estimator counts an image as 4800 chars; reuse that convention
@@ -84,7 +80,6 @@ function countUserContent(forecast: HistoryForecast, content: unknown, visionTar
   let inImageRun = false;
   for (const block of blockList(content)) {
     if (block.type === "image") {
-      forecast.imageCount += 1;
       if (visionTarget) forecast.imageChars += ESTIMATED_IMAGE_CHARS;
       else if (!inImageRun) forecast.imageChars += placeholderChars;
       inImageRun = true;
@@ -97,12 +92,10 @@ function countUserContent(forecast: HistoryForecast, content: unknown, visionTar
 
 function countToolCall(forecast: HistoryForecast, block: ForecastBlock, isSame: boolean): void {
   // Tool calls replay as id/name/arguments; cross-model the thoughtSignature is
-  // stripped, so it is never counted for a foreign target. Encrypted payload char
-  // length is the size proxy throughout (contextimate's convention).
+  // stripped, so it only counts for its own model. Encrypted payload char length is
+  // the size proxy throughout (contextimate's convention).
   forecast.textChars += JSON.stringify({ id: block.id, name: block.name, arguments: block.arguments }).length;
-  if (typeof block.thoughtSignature !== "string") return;
-  if (isSame) forecast.keptReasoningChars += block.thoughtSignature.length;
-  else forecast.droppedReasoningChars += block.thoughtSignature.length;
+  if (isSame && typeof block.thoughtSignature === "string") forecast.keptReasoningChars += block.thoughtSignature.length;
 }
 
 function countThinking(forecast: HistoryForecast, block: ForecastBlock, isSame: boolean): void {
@@ -110,9 +103,7 @@ function countThinking(forecast: HistoryForecast, block: ForecastBlock, isSame: 
   const readable = block.thinking ?? "";
   if (block.redacted) {
     // Redacted thinking is opaque encrypted content: replayed same-model, dropped otherwise.
-    if (signature === undefined) return;
-    if (isSame) forecast.keptReasoningChars += signature.length;
-    else forecast.droppedReasoningChars += signature.length;
+    if (isSame && signature !== undefined) forecast.keptReasoningChars += signature.length;
     return;
   }
   if (isSame) {
@@ -126,7 +117,6 @@ function countThinking(forecast: HistoryForecast, block: ForecastBlock, isSame: 
   // Cross-model: readable thinking becomes ordinary assistant text (blank blocks
   // vanish); the encrypted payload, if any, is dropped.
   if (readable.trim().length > 0) forecast.textChars += readable.length;
-  if (signature !== undefined) forecast.droppedReasoningChars += signature.length;
 }
 
 /**
@@ -138,18 +128,10 @@ function countThinking(forecast: HistoryForecast, block: ForecastBlock, isSame: 
  */
 export function forecastHistoryForTarget(messages: readonly ForecastMessage[], target: TargetModel): HistoryForecast {
   const visionTarget = target.input?.includes("image") ?? true;
-  const forecast: HistoryForecast = {
-    textChars: 0,
-    keptReasoningChars: 0,
-    droppedReasoningChars: 0,
-    imageCount: 0,
-    imageChars: 0,
-    messageCount: 0,
-  };
+  const forecast: HistoryForecast = { textChars: 0, keptReasoningChars: 0, imageChars: 0 };
   for (const message of messages) {
     if (message.role === "assistant") {
       if (message.stopReason === "error" || message.stopReason === "aborted") continue;
-      forecast.messageCount += 1;
       const isSame = sameModel(message, target);
       for (const block of blockList(message.content)) {
         if (block.type === "thinking") countThinking(forecast, block, isSame);
@@ -158,7 +140,6 @@ export function forecastHistoryForTarget(messages: readonly ForecastMessage[], t
       }
       continue;
     }
-    forecast.messageCount += 1;
     const placeholderChars = message.role === "toolResult" ? NON_VISION_TOOL_PLACEHOLDER_CHARS : NON_VISION_USER_PLACEHOLDER_CHARS;
     countUserContent(forecast, message.content, visionTarget, placeholderChars);
   }
