+++ tests/cachemire/model-switch.test.ts
@@ -1,16 +1,18 @@
 // Model-switch forecasting (issue #57): the clock's target-currency states, the sized
-// break prediction, computeSwitchForecast wiring over real pi session machinery, and
-// the event flow from model_select through message_end.
+// break prediction, and computeSwitchForecast wiring over real pi session machinery.
+// The event flow lives in model-switch-events.test.ts.
 import { test } from "node:test";
 import assert from "node:assert/strict";
 
-import piCachemire, { internals } from "../../extensions/pi-cachemire/index.ts";
+import { internals } from "../../extensions/pi-cachemire/index.ts";
 import { computeSwitchForecast, type SwitchTarget } from "../../extensions/pi-cachemire/forecast.ts";
-import type { ExtensionAPI, SessionEntry } from "@earendil-works/pi-coding-agent";
+import type { SessionEntry } from "@earendil-works/pi-coding-agent";
 
 const { cacheClock, predictBreak, renderBreakingLine, restoreLineageSnapshots, withinWarmHorizon } = internals;
 
 const CONTRACT_5M = { kind: "contract", ttlMs: 5 * 60_000, source: "observed" } as const;
+// The window an openai-codex snapshot really carries (contract TTLs are anthropic's).
+const BAND = { kind: "band", softMs: 5 * 60_000, hardMs: 60 * 60_000 } as const;
 const RATES = { input: 15, output: 75, cacheRead: 1.5, cacheWrite: 18.75 };
 const MIN = 60_000;
 
@@ -45,11 +47,11 @@ test("clock: model switch forecasts in the target currency, always marked est", 
 
 test("clock: A\u2192B\u2192A switch-back defers to the target's own prior entry", () => {
   // The target model's own last billed call is inside its window: "cold" would overclaim.
-  const back = cacheClock({ now: MIN, lastRequestAt: 0, window: CONTRACT_5M, modelSwitched: true, switchForecast: { ...FORECAST, prior: { requestAt: 0, window: CONTRACT_5M } } });
+  const back = cacheClock({ now: MIN, lastRequestAt: 0, window: CONTRACT_5M, modelSwitched: true, switchForecast: { ...FORECAST, prior: { requestAt: 0, window: BAND } } });
   assert.equal(back.phase, "warm-unknown");
   assert.equal(back.text, "cache may still be warm \u00b7 last gpt-5.6-luna call 1m ago \u00b7 next send confirms");
-  // Past the prior's window the switch-back is cold expected like any other switch.
-  const backCold = cacheClock({ now: 10 * MIN, lastRequestAt: 0, window: CONTRACT_5M, modelSwitched: true, switchForecast: { ...FORECAST, prior: { requestAt: 0, window: CONTRACT_5M } } });
+  // Past the prior's soft horizon the switch-back is cold expected like any other switch.
+  const backCold = cacheClock({ now: 10 * MIN, lastRequestAt: 0, window: CONTRACT_5M, modelSwitched: true, switchForecast: { ...FORECAST, prior: { requestAt: 0, window: BAND } } });
   assert.equal(backCold.phase, "cold");
 });
 
@@ -190,10 +192,39 @@ test("computeSwitchForecast: a billed source anchor calibrates the estimate", ()
   const calibrated = computeSwitchForecast({ ...base, snapshots: [snapshot] });
   assert.equal(calibrated.estTokens, 4_500);
   assert.deepEqual(calibrated.breakdown, { anchorTokens: 2_925, droppedThinking: 0 });
+  // A pre-api snapshot (older cachemire state) still anchors: the tokenizer is the
+  // model's, not the wire's.
+  assert.equal(computeSwitchForecast({ ...base, snapshots: [{ ...snapshot, api: undefined }] }).estTokens, 4_500);
   // Without a billed anchor on the path the heuristic number stands alone, unexplained.
   const plain = computeSwitchForecast({ ...base, snapshots: [] });
   assert.equal(plain.estTokens, 3_000);
   assert.equal(plain.breakdown, undefined);
+});
+
+test("computeSwitchForecast: a switch-back prior needs an exact api and an uncompacted path", () => {
+  const sol: SwitchTarget = { provider: "openai-codex", id: "gpt-5.6-sol", api: "openai-codex-responses" };
+  const snapshot = {
+    requestLeafId: "u1", responseEntryId: "a1", responseAt: 5_000, requestAt: 1_000,
+    promptTokens: 80_000, cacheRead: 0, cacheWrite: 0,
+    provider: "openai-codex", model: "gpt-5.6-sol", api: "openai-codex-responses", window: BAND,
+  };
+  const base = { target: sol, entries: entriesFixture(), activeLeafId: "a1", systemPromptChars: 0, tools: [] };
+  assert.deepEqual(
+    computeSwitchForecast({ ...base, snapshots: [snapshot] }).prior,
+    { requestAt: 1_000, window: BAND },
+  );
+  // Same id via a different (or unrecorded) wire API is a different cache: no warmth hint.
+  assert.equal(computeSwitchForecast({ ...base, snapshots: [{ ...snapshot, api: "openai-responses" }] }).prior, undefined);
+  assert.equal(computeSwitchForecast({ ...base, snapshots: [{ ...snapshot, api: undefined }] }).prior, undefined);
+  // A compaction between the prior call and the leaf rewrote the prefix it cached.
+  const compacted = [
+    ...entriesFixture(),
+    { type: "compaction", id: "c1", parentId: "a1", timestamp: "2026-07-01T10:01:00.000Z" },
+  ] as unknown as SessionEntry[];
+  assert.equal(
+    computeSwitchForecast({ ...base, entries: compacted, activeLeafId: "c1", snapshots: [snapshot] }).prior,
+    undefined,
+  );
 });
 
 test("computeSwitchForecast: gateway targets are labelled, not refused a number", () => {
@@ -209,130 +240,16 @@ test("computeSwitchForecast: gateway targets are labelled, not refused a number"
   assert.ok(forecast.estTokens !== undefined && forecast.estTokens > 0);
 });
 
-// --- event flow ---------------------------------------------------------------------------
-
-type Handler = (...args: unknown[]) => unknown;
-
-function extensionProbe(): { handlers: Map<string, Handler[]> } {
-  const handlers = new Map<string, Handler[]>();
-  const pi = {
-    on(event: string, handler: Handler): void {
-      handlers.set(event, [...(handlers.get(event) ?? []), handler]);
-    },
-    registerCommand(): void {},
-    getThinkingLevel: () => "off",
-    getActiveTools: () => [],
-    getAllTools: () => [],
-  } as unknown as ExtensionAPI;
-  piCachemire(pi);
-  return { handlers };
-}
-
-async function fire(probe: ReturnType<typeof extensionProbe>, event: string, ...args: unknown[]): Promise<void> {
-  for (const handler of probe.handlers.get(event) ?? []) await handler(...args);
-}
-
-function billedAssistant(model: string, api: string, prompt: number, userChars: number) {
-  return [
-    { type: "message", id: "u1", parentId: null, timestamp: "2026-07-01T10:00:00.000Z", message: { role: "user", content: "q".repeat(userChars), timestamp: 1_000 } },
-    { type: "message", id: "a1", parentId: "u1", timestamp: "2026-07-01T10:00:05.000Z", message: {
-      role: "assistant", content: [{ type: "text", text: "answer" }], provider: "anthropic", api, model,
-      stopReason: "stop", timestamp: Date.now(), usage: usage(2, prompt - 2, 0),
-    } },
-  ];
-}
-
-function probeContext(entries: unknown[], api: string, notifications: string[], widgets: string[]) {
-  return {
-    hasUI: true,
-    ui: {
-      theme: undefined,
-      setWidget(_key: string, widget?: unknown): void {
-        // Cachemire also registers a TUI-capture hook through setWidget; only record
-        // real widget lines.
-        if (typeof widget === "function") widget({ requestRender: () => {} });
-        else widgets.push(Array.isArray(widget) ? widget.join("\n") : "");
-      },
-      notify(text: string): void {
-        notifications.push(text);
-      },
-    },
-    model: {
-      id: "claude-opus-4-8", provider: "anthropic", api, reasoning: false, contextWindow: 200_000,
-      cost: { input: 15, output: 75, cacheRead: 1.5, cacheWrite: 18.75 },
-    },
-    sessionManager: { getEntries: () => entries, getLeafId: () => "a1" },
-    getSystemPrompt: () => "",
-  };
-}
-
-const ANTHROPIC_PAYLOAD = {
-  model: "claude-opus-4-8",
-  system: [{ type: "text", text: "fixture", cache_control: { type: "ephemeral" } }],
-  messages: [{ role: "user", content: [{ type: "text", text: "next" }] }],
-};
-
-test("event flow: the same model over a different wire API is a switch", async () => {
-  // Baseline billed over anthropic-messages; the session resumes on bedrock-anthropic.
-  const entries = billedAssistant("claude-opus-4-8", "anthropic-messages", 80_000, 200);
-  const notifications: string[] = [];
-  const widgets: string[] = [];
-  const probe = extensionProbe();
-  const ctx = probeContext(entries, "bedrock-anthropic", notifications, widgets);
-
-  await fire(probe, "session_start", {}, ctx);
-  try {
-    assert.match(widgets.at(-1)!, /model switched/, "an API-only switch must flip the clock at session_start");
-    assert.match(widgets.at(-1)!, /\(est\)/, "the forecast must be marked as an estimate");
-
-    // Send time: the tiny estimated re-write is below the materiality thresholds, so
-    // no break notice is posted (the widget already explains the switch).
-    await fire(probe, "before_provider_request", { payload: ANTHROPIC_PAYLOAD }, ctx);
-    assert.equal(notifications.length, 0, "an immaterial estimated re-write must not post a notice");
-    assert.match(widgets.at(-1)!, /model switched/);
-
-    // Usage billed over the new API re-baselines: the switch state clears.
-    await fire(probe, "message_end", { message: {
-      role: "assistant", content: [], provider: "anthropic", api: "bedrock-anthropic", model: "claude-opus-4-8",
-      stopReason: "stop", timestamp: Date.now(), usage: usage(80_000, 0, 0),
-    } });
-    assert.doesNotMatch(widgets.at(-1)!, /model switched/, "fresh usage must re-baseline the currency");
-  } finally {
-    await fire(probe, "session_shutdown", {});
-  }
-});
-
-test("event flow: a material estimated re-write posts one est-marked notice", async () => {
-  // 200k chars of history, and the same id billed 80k through the other api: with
-  // identical heuristics on both sides, the density calibration anchors the estimate
-  // to the billed prompt exactly. Still far above the 20k threshold.
-  const entries = billedAssistant("claude-opus-4-8", "anthropic-messages", 80_000, 200_000);
-  const notifications: string[] = [];
-  const widgets: string[] = [];
-  const probe = extensionProbe();
-  const ctx = probeContext(entries, "bedrock-anthropic", notifications, widgets);
-
-  await fire(probe, "session_start", {}, ctx);
-  try {
-    await fire(probe, "before_provider_request", { payload: ANTHROPIC_PAYLOAD }, ctx);
-    assert.equal(notifications.length, 1, "a material estimated re-write must post a notice");
-    assert.match(notifications[0]!, /sending ~80\.0k uncached to anthropic \(est/);
-    assert.match(notifications[0]!, /model switched/);
-  } finally {
-    await fire(probe, "session_shutdown", {});
-  }
-});
-
 test("computeSwitchForecast: finds the target's own prior call on the active path only", () => {
   const entries = entriesFixture();
-  const snapshots = restoreLineageSnapshots(entries, () => CONTRACT_5M);
+  const snapshots = restoreLineageSnapshots(entries, () => BAND);
   // Switching back to sol itself: its billed call at a1 is on the path.
   const backToSol = computeSwitchForecast({
     target: { provider: "openai-codex", id: "gpt-5.6-sol", api: "openai-codex-responses" },
     entries, activeLeafId: "a1", systemPromptChars: 0, tools: [], snapshots,
   });
   assert.ok(backToSol.prior, "the target's own billed call must be found");
-  assert.equal(backToSol.prior!.window!.kind, "contract");
+  assert.equal(backToSol.prior!.window!.kind, "band");
   // Switching to a model that never billed on this path: no prior.
   const toOpus = computeSwitchForecast({
     target: OPUS, entries, activeLeafId: "a1", systemPromptChars: 0, tools: [], snapshots,
